+++ packages/composites/market-closure/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @chainlink/market-closure-adapter
 
+## 1.3.61
+
+### Patch Changes
+
+- Updated dependencies [[`41f2fa5`](https://github.com/smartcontractkit/external-adapters-js/commit/41f2fa5c81aec7416436c9f9e31ced15e307af7f)]:
+  - @chainlink/finnhub-adapter@3.3.1
+
+## 1.3.60
+
+### Patch Changes
+
+- Updated dependencies [[`7b916bf`](https://github.com/smartcontractkit/external-adapters-js/commit/7b916bf6ab3b69daa0dbb5e043265cc0fd1a262b)]:
+  - @chainlink/finnhub-adapter@3.3.0
+
+## 1.3.59
+
+### Patch Changes
+
+- Updated dependencies [[`d5bef44`](https://github.com/smartcontractkit/external-adapters-js/commit/d5bef440cf0540ef9199382bc1f7129daedf3021)]:
+  - @chainlink/finnhub-adapter@3.2.1
+
 ## 1.3.58
 
 ### Patch Changes